requirements.txt
Django==4.2.10
django-ninja==1.1.0
django-cors-headers
python-dotenv
psycopg2-binary
requests
pandas
numpy
openpyxl
beautifulsoup4
fuzzywuzzy
python-Levenshtein
openai
whitenoise
pydantic

djangorestframework
urllib3
duckdb>=0.9.2
pgvector>=0.2.4
django-redis
redis
